Move Argv type to it's own file
